# src/main/resources/api.yaml
openapi: '3.0.3'
info:
  description: |
    Prueba tecnica
  version: 1.0.0
  x-template-version: 3.0.0
  title: Prueba tecnica
  contact:
    name: Jhonatan Acuña Hernández
    email: jonathan.961120@hotmail.com
    url: 'https://www.linkedin.com/in/jhonaa96/'

servers:
    - url: '/api/v1'

paths:
  /characters/getCharacters:
    get:
      tags:
        - characters
      summary: Get all characters
      operationId: characters
      responses:
          '200':
            description: 'Characters'
            content:
                application/json:
                  schema:
                      $ref: '#/components/schemas/Characters'
  /characters/saveCharacter:
    post:
      tags:
        - characters
      summary: save character
      operationId: saveCharacter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
      responses:
        '200':
          description: 'Character'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

components:
  schemas:
    Response:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Data'
        notification:
          $ref: '#/components/schemas/Notification'
    Data:
      type: object

    Notification:
      type: object
      properties:
        category:
          type: string
        code:
          type: string
        message:
          type: string
        description:
          type: string

    Characters:
      type: array
      items:
        $ref: '#/components/schemas/Character'
    Character:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        gender:
          type: string
        image:
          type: string
    Result:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        species:
          type: string
        type:
          type: string
        gender:
          type: string
        origin:
          type: object
        location:
          type: object
        image:
          type: string
        episode:
          type: array
          items:
            type: string
        url:
          type: string
        created:
          type: string